backend/requirements.txt
streamlit
supabase
requests
flask
python-dotenv
pytest
transformers
PyPDF2
intasend-python
python-docx
fastapi
uvicorn[standard]
pydantic
huggingface-hub